bug fix #2619
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7067 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/src/org/wcs/smart/i2/xml/XmlToIntelData.java
@@ -246,7 +246,7 @@ public class XmlToIntelData {
 				if (!Files.exists(trg.getParent())) {
 					Files.createDirectories(trg.getParent());
 				}
-				Files.copy(src, trg, StandardCopyOption.REPLACE_EXISTING);
+				if (Files.exists(src)) Files.copy(src, trg, StandardCopyOption.REPLACE_EXISTING);
 			} catch (IOException e1) {
 				copyErrors.add(Messages.XmlToIntelData_SourceCopyError);
 				Intelligence2PlugIn.log(e1.getMessage(), e1);
